+++ src/ReadTask.cpp
@@ -57,7 +57,7 @@ bool ReadTask::run()
     // todo: do the reading and writing in chunks
     // todo: need to make this use O_DIRECT on the IOC side
     ioc->willRead(cmd->filename, cmd->offset, cmd->count);
-    int err;
+    ssize_t err;
     while ((uint) resp->returnCode < cmd->count)
     {
         try
